Move Abstract into root package
